simpleBase [turtlBase] release
simpleBase [a.k.a turtlBase] is base for LiquidBounce's JavaScript API which tries to automatically recognize and fix the errors. It's great for begginers and not only for them  (imo). Totally free to use and you can publish your script using it without any problems (it would be cool if you just type that you used that base)